Update README.md
This is a very simple project but with some important topics approached where we used React Hook and Styled Components to bring some styling to the project. The app is just a counter with a function to Increment and Decrement the values on the counter. To achiave the expected result we used React Hook useState where we're able to call this funcionality in order to change the state of the app over and over incrementing and decrenenting the values. I hope you like.